requirements.txt
yfinance>=0.2.36
gspread>=6.0.0
google-auth>=2.27.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
